+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.4] - 2026-09-11
+
 ### Changed
 
 - `QH9PaddedSource.get_batch_at` is renamed `read_batch(start, size)`. It reads molecules on